// src/melee/types.ts
export enum Character {
  CAPTAIN_FALCON = 0,
  DONKEY_KONG = 1,
  FOX = 2,
  GAME_AND_WATCH = 3,
  KIRBY = 4,
  BOWSER = 5,
  LINK = 6,
  LUIGI = 7,
  MARIO = 8,
  MARTH = 9,
  MEWTWO = 10,
  NESS = 11,
  PEACH = 12,
  PIKACHU = 13,
  ICE_CLIMBERS = 14,
  JIGGLYPUFF = 15,
  SAMUS = 16,
  YOSHI = 17,
  ZELDA = 18,
  SHEIK = 19,
  FALCO = 20,
  YOUNG_LINK = 21,
  DR_MARIO = 22,
  ROY = 23,
  PICHU = 24,
  GANONDORF = 25,
}

export enum Stage {
  FOUNTAIN_OF_DREAMS = 2,
  POKEMON_STADIUM = 3,
  PEACHS_CASTLE = 4,
  KONGO_JUNGLE = 5,
  BRINSTAR = 6,
  CORNERIA = 7,
  YOSHIS_STORY = 8,
  ONETT = 9,
  MUTE_CITY = 10,
  RAINBOW_CRUISE = 11,
  JUNGLE_JAPES = 12,
  GREAT_BAY = 13,
  HYRULE_TEMPLE = 14,
  BRINSTAR_DEPTHS = 15,
  YOSHIS_ISLAND = 16,
  GREEN_GREENS = 17,
  FOURSIDE = 18,
  MUSHROOM_KINGDOM = 19,
  MUSHROOM_KINGDOM_2 = 20,
  VENOM = 22,
  POKE_FLOATS = 23,
  BIG_BLUE = 24,
  ICICLE_MOUNTAIN = 25,
  ICETOP = 26,
  FLAT_ZONE = 27,
  DREAMLAND = 28,
  YOSHIS_ISLAND_N64 = 29,
  KONGO_JUNGLE_N64 = 30,
  BATTLEFIELD = 31,
  FINAL_DESTINATION = 32,
}